requirements.txt
google-api-python-client==2.111.0
groq==0.4.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
